feat(db): seeded condenser account - with Unique name constraint
