+++ cmd/main.go
@@ -7,21 +7,25 @@ import (
 	"mimic/modules/db"
 	"mimic/modules/db/mimic"
 	"mimic/modules/db/mimic/blocks"
+	"mimic/modules/db/mimic/condenserdb"
 	"mimic/modules/db/mimic/state"
 )
 
 func main() {
 
 	db := db.New(db.NewDbConfig())
+	db.Init()
 	mimicDb := mimic.New(db)
+	mimicDb.Init()
+
 	hiveBlocks := blocks.New(mimicDb)
 	stateDb := state.New(mimicDb)
+	condenserDb := condenserdb.New(mimicDb)
 
 	plugins := []aggregate.Plugin{
-		db,
-		mimicDb,
 		hiveBlocks,
 		stateDb,
+		condenserDb,
 	}
 
 	agg := aggregate.New(plugins)
